+++ src/routes/stock/in.tsx
@@ -1,9 +1,0 @@
-import { SpokeLayout } from "@/components/layout/SpokeLayout";
-import Stub from "@/components/layout/Stub";
-export default function StockIn() {
-  return (
-    <SpokeLayout title="Stock in">
-      <Stub name="Stock in" />
-    </SpokeLayout>
-  );
-}